2024-12-09 install.sh - master branch
This PR follows on from discussion in #651 at
[@6ffm70](https://github.com/SensorsIot/IOTstack/issues/651#issuecomment-2524083300)
post and later.

Previously, the `backups` and `services` folders were only created if
the `~/IOTstack` folder had just been cloned.

As well as "installing", the install script should also function as a
"repair" script so it is preferable to always create the `backups` and
`services` directories and ensure that they, and their contents,
have correct ownership.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>