2025-03-16 Pi-hole - master branch
A DM on the IOTstack Discord channel pointed out that the Pi-hole
documentation did not explain how to alter the resolver configuration
in the presence of NetworkManager.

This PR adds the necessary instructions.

Takes the opportunity to move the focus away from an assumption of a
Raspberry Pi (and discussions of Bullseye-and-earlier vs Bookworm; or
Raspberry Pi OS vs Debian) by adopting the term "Pi-hole system" as
meaning "The host platform where the Pi-hole service is running."

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>